+++ whatsnew.md
@@ -1,4 +1,4 @@
 # This file will provide extra info on new updates / commits.
 
 # WHAT'S NEW [23/12/2025]
-  - Added PEB Module Loader structs
+  - Added Shutdown fields to PEB